clipboard-apis/readText.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Clipboard API: navigator.clipboard.readText() - negative test</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <h1>Negative Test for navigator.clipboard.readText() without permission</h1>

  <script>
    promise_test(async t => {
      await navigator.permissions.revoke({ name: "clipboard-read" }); // ensure permission is not granted
      try {
        await navigator.clipboard.readText();
        assert_unreached("readText() should not resolve without permission");
      } catch (e) {
        assert_equals(e.name, "NotAllowedError", "Should throw NotAllowedError without clipboard permission");
      }
    }, "navigator.clipboard.readText() throws NotAllowedError without permission");
  </script>
</body>
